+++ packages/hps-srv-mock/src/load-config/load-mock-config-file.ts
@@ -11,7 +11,7 @@ export const loadMockConfigFile = async (
   configLoaderOptions: ConfigLoaderOptions = {
     configFile: configFileName,
     loaderOptions: {
-      externals: [/^@hyperse\/.*/],
+      externals: [/^@hyperse\/.*/, /^@hyperse-hub\/.*/],
     },
   }
 ): Promise<DeepPartial<HpsMockOptions>> => {